api/ml/requirements.txt
sentence-transformers
numpy
scipy
scikit-learn
joblib
hdbscan
faiss-cpu
fastapi
uvicorn
umap-learn
asyncio

